Updates galore!
Fixed purge command, added volume command. Also added a reload {extension} command to save having to spam unload and reload, moved audio files to a voice folder and updated the code in recognition of this change. Finally, I've changed quite a few of the bot's responses to embeds (aesthetics) and removed the 'cats.mp3' file and command. You can see why, and I'm disappointed in myself for even listening to my friend when he said to add it.