Quite a few updates
Added administration module in which you can enable/disable the welcome module, slur detector, and music module.

You can also set up the slur detector and music modules for your guild.

Server info is now stored in a JSON file which is much simpler than an SQLite database.

This means that server administrators can now set the bot's prefix within the guild to avoid conflicts with other bots.

Bot on_ready() event moved to the 'owner' module to allow for the new, periodically changing bot status.

Moved the welcomesetup command to the administration module.

Note: the slur detector module is broken by this update - the bot no longer responds to reactions in the moderation channel. I need to take a deeper look into it but wanted to get this update out now as it is rather significant.

I will remove the deprecated databases in the next commit, but wanted to back them up here first.