Quick fixes and new commands
Removed TechSMP ad from help command.

Added 'echo' command to owner module (bot repeats message).

Fixed mention for prefix, as it was sending it for @here and @everyone, as well as with messages from bots.

Moved guildSetup to new events module that sends a log message to the TechBot server when the bot is added to or removed from a server.

Some slur detector fixes - not perfect yet, though.