Sunset Slur Detector Module
It's time to say goodbye to the slur detector module. I liked the idea and genuinely wanted to use it, but couldn't get it working and so sidelined it to kind of be forgotten.

Now, Discord has added moderation tools of its own that allow server admins to do essentially the same thing. Therefore, it doesn't make sense for me to keep working on it, especially with Discord locking down permissions on message content access and focusing on slash commands.

The good news is that this should probably improve the bot's performance, since every message no longer has to be checked for slurs. Discord can do that now (as they should have been for a long time!).